Add support for scroll(1)
Scroll is a program that store all the lines of its child and can
be used in st as a way of implementing scrollback.

This solution is much better then implementing the scrollback in st
itself because having a different program allows to use it in any other
program without doing modification to those programs.

This commit comes from git://git.suckless.org/st.